refactor: rename module from cyberark_mcp to mcp_privilege_cloud
## Overview
Complete module rename to align with project naming conventions:
- Package: mcp-privilege-cloud
- Module: mcp_privilege_cloud (was: cyberark_mcp)
- Script: mcp-privilege-cloud

## Changes Made

### Module Structure
- Renamed src/cyberark_mcp/ → src/mcp_privilege_cloud/
- Updated all Python imports across codebase
- Updated pyproject.toml package configuration

### Documentation Updates
- Updated README.md, CLAUDE.md, DEVELOPMENT.md
- Updated docs/TESTING.md, docs/TROUBLESHOOTING.md
- Updated all Serena memory files
- Updated import examples and module references

### Configuration Updates
- Updated pyproject.toml package discovery and entry points
- Updated test coverage paths and tool configurations
- Updated CI/CD pipeline configuration
- Updated debug and utility scripts

### Testing
- ✅ All 175 tests pass with new module structure
- ✅ Module imports work correctly
- ✅ CLI execution verified: python -m mcp_privilege_cloud

## Breaking Changes
- Import statements must be updated: cyberark_mcp → mcp_privilege_cloud
- Module execution command changed: python -m cyberark_mcp → python -m mcp_privilege_cloud

## Verification
- Zero remaining cyberark_mcp references
- Full test suite passes (175/175)
- Documentation consistency verified
