+++ .serena/memories/config_centralization_implementation_details.md
@@ -1,0 +1,206 @@
+# Configuration Centralization - Implementation Details
+
+## Timeline and Dependencies
+
+### Week 1: Core Foundation
+- **Days 1-2**: Subagent 1 completes core configuration system
+- **Days 3-4**: Subagent 2 migrates authentication 
+- **Day 5**: Integration testing and validation
+
+### Week 2: Server Integration  
+- **Days 1-2**: Subagent 3 migrates server configuration
+- **Days 3-4**: Subagent 4 completes MCP integration
+- **Day 5**: Comprehensive testing and documentation
+
+### Week 3: Validation and Cleanup
+- **Days 1-2**: Full integration testing across all scenarios
+- **Days 3-4**: Documentation updates and examples
+- **Day 5**: Final validation and deployment preparation
+
+### Dependency Management
+```
+Subagent 1 (Config System) 
+    ↓
+Subagent 2 (Auth Migration) ← Subagent 3 (Server Migration)
+    ↓                            ↓
+    └─────── Subagent 4 (Integration) ──────┘
+```
+
+## Risk Assessment
+
+### Technical Risks
+
+| Risk | Probability | Impact | Mitigation Strategy |
+|------|-------------|--------|-------------------|
+| **Backward Compatibility Break** | Low | High | Maintain all existing `from_environment()` methods; comprehensive regression testing |
+| **Configuration Validation Issues** | Medium | Medium | Extensive validation testing; clear error messages; fallback to defaults |
+| **Environment Variable Conflicts** | Low | Low | Careful testing across different deployment scenarios |
+| **Test Suite Regression** | Low | High | Run full test suite after each migration phase |
+
+### Operational Risks
+
+| Risk | Probability | Impact | Mitigation Strategy |
+|------|-------------|--------|-------------------|
+| **Deployment Configuration Errors** | Medium | Medium | Enhanced error messages; configuration validation at startup |
+| **Documentation Outdated** | High | Low | Update documentation as part of each subagent task |
+| **Developer Confusion** | Low | Low | Clear migration guide; maintain old patterns during transition |
+
+## Success Criteria
+
+### Technical Success Criteria
+- [ ] All 124+ existing tests pass without modification
+- [ ] New configuration system has 100% test coverage
+- [ ] All environment variables load through centralized `ConfigManager`
+- [ ] Configuration validation provides clear, actionable error messages
+- [ ] No code duplication in environment variable handling
+
+### Quality Success Criteria  
+- [ ] PEP 8 compliance maintained across all new code
+- [ ] Type hints and docstrings for all new public methods
+- [ ] Comprehensive error handling with logging
+- [ ] Security best practices maintained (no credential logging)
+
+### User Experience Success Criteria
+- [ ] No changes to existing deployment procedures
+- [ ] Improved error messages for configuration issues
+- [ ] Enhanced debugging capabilities for configuration problems
+- [ ] Clear documentation for all configuration options
+
+### Performance Success Criteria
+- [ ] No degradation in application startup time
+- [ ] Configuration caching reduces repeated environment variable access
+- [ ] Memory usage remains stable with configuration caching
+
+## Testing Strategy
+
+### Unit Testing Approach
+1. **Configuration Tests** (`tests/test_config.py`)
+   - Environment variable loading and validation
+   - Default value handling
+   - Error conditions and edge cases
+   - Configuration caching behavior
+
+2. **Integration Tests** (existing test files)
+   - All existing tests continue to pass
+   - New `from_config()` methods tested
+   - Configuration error handling in realistic scenarios
+
+3. **Test Utilities**
+   - `ConfigTestUtils.create_test_config()` for easy test configuration
+   - Environment variable mocking utilities
+   - Configuration reset utilities for test isolation
+
+### Testing Validation Approach
+```bash
+# Run all tests to ensure no regression
+pytest                                    # All 124+ tests pass
+
+# Run specific configuration tests  
+pytest tests/test_config.py               # New configuration tests
+
+# Run integration tests with new config
+pytest tests/test_integration.py          # Integration tests
+
+# Test with various environment configurations
+CYBERARK_LOG_LEVEL=DEBUG pytest          # Test with different log levels
+CYBERARK_API_TIMEOUT=60 pytest           # Test with different timeouts
+```
+
+## Migration Strategy
+
+### Phase 1: Foundation (Week 1)
+1. **Create Core Configuration System** (Subagent 1)
+   - Implement complete `config.py` module
+   - Add comprehensive test coverage
+   - Validate all environment variable handling
+   - No changes to existing code
+
+### Phase 2: Authentication Migration (Week 1-2) 
+1. **Migrate Authentication** (Subagent 2)
+   - Add `from_config()` method to `CyberArkAuthenticator`
+   - Update `from_environment()` to use `ConfigManager`
+   - Ensure all existing tests pass unchanged
+   - Validate authentication functionality
+
+### Phase 3: Server Migration (Week 2)
+1. **Migrate Server** (Subagent 3)  
+   - Add `from_config()` method to `CyberArkMCPServer`
+   - Update server initialization logic
+   - Maintain backward compatibility
+   - Validate all server functionality
+
+### Phase 4: Complete Integration (Week 2-3)
+1. **MCP Integration** (Subagent 4)
+   - Update MCP server logging configuration
+   - Migrate integration tests
+   - Add configuration documentation
+   - Final validation and cleanup
+
+### Rollback Strategy
+- All changes maintain backward compatibility
+- Existing `from_environment()` methods continue to work
+- Configuration can be reverted by removing imports from new config module
+- No database or persistent state changes
+
+## Deliverables by Subagent
+
+### Subagent 1 Deliverables
+- `src/mcp_privilege_cloud/config.py` - Core configuration module (300+ lines)
+- `tests/test_config.py` - Comprehensive configuration tests (25+ tests)
+- Configuration validation framework with clear error messages
+- Testing utilities for other subagents
+
+### Subagent 2 Deliverables  
+- Updated `src/mcp_privilege_cloud/auth.py` with `from_config()` method
+- Migrated `from_environment()` to use `ConfigManager`
+- Updated authentication tests to work with new system
+- Backward compatibility maintained
+
+### Subagent 3 Deliverables
+- Updated `src/mcp_privilege_cloud/server.py` with `from_config()` method  
+- Migrated server initialization to centralized configuration
+- Updated server tests to work with new system
+- Enhanced error handling for configuration issues
+
+### Subagent 4 Deliverables
+- Updated `src/mcp_privilege_cloud/mcp_server.py` logging configuration
+- Migrated all integration tests to new configuration system
+- Configuration troubleshooting documentation
+- Final integration validation and cleanup
+
+## Integration Points
+
+### Configuration Loading Flow
+```python
+# New centralized flow:
+ConfigManager.get_config() 
+    → Load from environment variables
+    → Validate all configuration 
+    → Cache validated configuration
+    → Return immutable CyberArkConfig object
+
+# Classes use configuration:
+CyberArkAuthenticator.from_config(config)
+CyberArkMCPServer.from_config(config)
+```
+
+### Error Handling Integration
+```python
+# Enhanced error messages:
+ValueError: Configuration validation failed:
+  - CYBERARK_IDENTITY_TENANT_ID is required
+  - CYBERARK_API_TIMEOUT must be between 5 and 300 seconds
+  
+# Clear troubleshooting guidance included
+```
+
+### Testing Integration
+```python
+# Test utilities available across all test files:
+config = ConfigTestUtils.create_test_config(api_timeout=60)
+authenticator = CyberArkAuthenticator.from_config(config)
+
+# Environment mocking:
+with ConfigTestUtils.mock_environment({'CYBERARK_LOG_LEVEL': 'DEBUG'}):
+    config = ConfigManager.get_config()
+```
