+++ .serena/memories/configuration_centralization_plan.md
@@ -1,0 +1,38 @@
+# Configuration Centralization Implementation Plan
+
+**Project**: CyberArk Privilege Cloud MCP Server  
+**Opportunity**: #4 - Configuration Centralization  
+**Version**: 1.0  
+**Date**: June 2025  
+
+## Executive Summary
+
+### Objective
+Transform the current scattered configuration management across multiple modules into a centralized, type-safe, and maintainable configuration system.
+
+### Business Impact
+- **Reduced maintenance overhead**: Eliminate duplicate environment variable validation
+- **Improved developer experience**: Single source of truth for all configuration
+- **Enhanced reliability**: Comprehensive validation with clear error messages
+- **Better testability**: Centralized configuration utilities for testing
+
+### Current Issues
+- Configuration logic scattered across `auth.py`, `server.py`, and `mcp_server.py`
+- Duplicate environment variable handling (`CYBERARK_API_TIMEOUT` in multiple places)
+- Hardcoded defaults and magic numbers throughout codebase
+- No comprehensive validation framework
+- Complex test mocking for each `from_environment()` method
+
+### Proposed Solution
+Centralized `ConfigManager` with type-safe configuration handling, comprehensive validation, and unified testing utilities while maintaining 100% backward compatibility.
+
+## Implementation Strategy
+
+### Subagent Task Breakdown
+1. **Core Configuration System** - Create centralized config infrastructure
+2. **Authentication Migration** - Migrate auth.py to use centralized config
+3. **Server Migration** - Migrate server.py to use centralized config  
+4. **MCP Integration** - Update mcp_server.py and testing utilities
+
+### Timeline: 2 weeks with parallel execution
+### Risk: Low - maintains backward compatibility throughout
