+++ .serena/memories/mcp_server_standardization_solution.md
@@ -1,0 +1,63 @@
+# MCP Server Standardization Solution
+
+## Problem Analysis
+Currently the MCP server has multiple execution methods:
+1. `python run_server.py` (recommended multiplatform launcher)
+2. `python src/mcp_privilege_cloud/mcp_server.py` (direct execution)
+3. `python -m src.mcp_privilege_cloud.mcp_server` (module execution)
+
+Issues with current approach:
+- No proper Python packaging (missing `pyproject.toml`)
+- Manual path manipulation in `run_server.py` 
+- Not compatible with modern Python tooling (`uv`/`uvx`)
+- Multiple confusing entry points
+- Uses `requirements.txt` instead of modern dependency management
+
+## Proposed Standardized Solution
+
+### Single Entry Point Strategy
+**Primary Method:** `uvx mcp-privilege-cloud`
+- Works for external users without local installation
+- Automatically handles dependencies
+- Modern Python tooling standard
+
+**Development Method:** `uv run mcp-privilege-cloud`
+- For local development
+- Uses project dependencies
+
+**Fallback Method:** `python -m mcp_privilege_cloud`
+- For environments without `uv`
+- Maintains backward compatibility
+
+### Technical Implementation
+
+1. **Create `pyproject.toml`** with proper packaging configuration
+   - Project metadata and dependencies
+   - Console script entry point: `mcp-privilege-cloud = mcp_privilege_cloud:main`
+   - Tool configuration (pytest, ruff, etc.)
+
+2. **Enhance `src/mcp_privilege_cloud/__init__.py`**
+   - Add `main()` function as primary entry point
+   - Import and delegate to `mcp_server.main()`
+   - Handle environment setup and error handling
+
+3. **Update `src/mcp_privilege_cloud/__main__.py`**
+   - Enable `python -m mcp_privilege_cloud` execution
+   - Simple delegation to main entry point
+
+4. **Deprecate `run_server.py`**
+   - Keep for backward compatibility
+   - Add deprecation warning
+   - Redirect to new entry point
+
+5. **Migration from `requirements.txt` to `pyproject.toml`**
+   - Move all dependencies to `pyproject.toml`
+   - Keep `requirements.txt` for backward compatibility
+
+### Benefits
+- Single, clear execution method
+- Modern Python packaging standards
+- `uv`/`uvx` compatibility
+- Cleaner project structure
+- Better dependency management
+- Maintains backward compatibility
